backend/requirements.txt
flask==3.0.2
python-dotenv==1.0.0
ultralytics==8.0.0
opencv-python==4.9.0.80
pymongo==4.6.1
eventlet==0.33.3